Refactor race form into partial class
